gst/elements/: fix some signedness
Original commit message from CVS:

* gst/elements/gstfakesrc.h:
* gst/elements/gstfilesrc.c: (gst_file_src_create_read):
* gst/elements/gsttypefindelement.c:
fix some signedness
* gst/elements/gstfilesink.c: (gst_file_sink_render):
I wonder if this could actually write +2GB files before

+++ plugins/elements/gstfilesrc.c
@@ -36,6 +36,7 @@
 #include "gstfilesrc.h"
 
 #include <stdio.h>
+#include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
 #ifdef HAVE_UNISTD_H
@@ -731,7 +732,7 @@ gst_file_src_create_read (GstFileSrc * src, guint64 offset, guint length,
     goto could_not_read;
 
   /* regular files should have given us what we expected */
-  if (ret < length && src->is_regular)
+  if ((gint) ret < length && src->is_regular)
     goto unexpected_eos;
 
   /* other files should eos if they read 0 */
